backend/ingestion/docs/requirements.txt
pymupdf==1.22.5
ebooklib==0.18
python-dotenv==1.0.1
